move model export to integration, tidy up float formattting
